Enhances the user experience and adds new resources
Improves the website's usability by adding a "Messages" link to the navigation, which directs users to the podcast on Spotify.

Adds an "Announcements" section to the homepage to communicate important updates.

Introduces a section featuring more books on the "Books" page, directing users to the external website.

Updates ministry logos on the ministry page.

Removes the slug field from the Book admin interface.